fix: address PR review feedback (CHANGELOG scope, truncateOutput dedup, entry text)
- Restore 58 historical CHANGELOG entries deleted outside spec scope
- Rewrite CHANGELOG entry to describe user-visible change (DR-003)
- Extract shared TruncateOutput to internal/textutil (CS-004)
- Remove duplicate truncateOutput from scaffold and setup packages
- Move truncateOutput tests to internal/textutil/output_test.go
- Update AGENTS.md project structure to include textutil/

Addresses review feedback from @jflowers on PR #269.
